Move the Json parsing to the new component
